Make sam respect the XDG Base Directory specification
One of the few XDG things I unironically like - it's nice to not have
a bunch of littering in my home directory.

This does a few things:

- All sam files live in a subdirectory now.
- samrc now lives in XDG_CONFIG_HOME, which by default is ~/.config/
- sam.save & sam.exit both live in XDG_STATE_HOME, default
  ~/.local/state
- XDG_STATE_HOME is automatically created if it doesn't exist.

There's still some files that are created in the home
directory (anything to do with the sockets), I'll need to play around
with those to see if I want to put them in XDG_STATE_HOME as well, but
honestly those files are deleted on normal exit, and you should be
able to work out that code by yourself if that really matters to you.

Closes #4.