Fix RMB scroll behavior
Partial fix for #3 - the RMB will now not allow you to scroll beyond
the end of the buffer.  This may break in some edge cases, e.g. having
a frame be unreasonably small, but that will require testing to
determine - this is a fine fix for the general case for now.