Purge wchar_t and replace with Rune
Ironically I think this is reverting an earlier refactor, lol.

wchar_t may be in the standard, but it sucks. utf16 is a bad way
to encode Unicode - it's the worst of both worlds (bigger than UTF-8,
but with none of the guarantees of UCS-4/UTF-32) - and in many cases
bad implementations prevent it from supporting unicode, outside of
the BMP.  UTF-8 is the standard now; if a program *needs* to store
things in "one rune per code-point" style like sam does, it's best
to just go straight to UCS-4.