Bounds check when moving up a line
I'm not sure if the change in sam proper is needed (it just affirms
that dot has moved), but the one in samterm definitely is. Previously,
it was possible to move in an essentially random walk by going to EoF
and then moving up one line. Presumably nobody got around to fixing
this because... sam historically mostly supported the mouse.

May address #14 but I still have to test @aksr's cases.