Get libthread compiling
Expected this to be easier, but it ended up being actually pretty
complicated. There's a lot of stuff that was in libc.h that I had to
choose a new home for, and I had to mess with some platform-specific
stuff. So, watch this space.  Thankfully, I have access to a Mac and
an OpenBSD machine so I should be able to test this when we get closer
to done.

Removed iorw.c as AFAICT it's not called anywhere.